use interface for GameBoard
