+++ java/src/main/java/com/_98point6/droptoken/dto/game/Move.java
@@ -9,6 +9,4 @@ public class Move {
     private final String moveType;
     private final String player;
     private final int column;
-
-    //TODO to_json, exclude column based on moveType
 }
